Update firefoxoverload.py
This is a Python script I wrote so that I could easily execute the two scripts within it simultaneously - which is a simple task in Python. The Bash script does the integral work relating to Firefox overload and the eatmemory script is something in C written by: Jullman99. This guy is good at what he does. 

Here is the original C repo for Jullman: https://github.com/julman99/eatmemory